+++ realtec_vision_buddmeyer/ui/widgets/__init__.py
@@ -8,7 +8,6 @@ from .status_panel import StatusPanel
 from .metrics_chart import MetricsChart
 from .event_console import EventConsole
 from .log_viewer import LogViewer
-from .gentl_camera_settings_dialog import GenTLCameraSettingsDialog
 
 __all__ = [
     "VideoWidget",
@@ -16,5 +15,4 @@ __all__ = [
     "MetricsChart",
     "EventConsole",
     "LogViewer",
-    "GenTLCameraSettingsDialog",
 ]